% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertParameters.R
\name{convertParameters}
\alias{convertParameters}
\title{Parameter estimates from a nonlinear lest squares (NLS) fit.}
\usage{
convertParameters(curve)
}
\arguments{
\item{curve}{A fitted model object from nls function. The fitted model should have the following
parameter estimates - amplitude, intercept, frequency, phase.}
}
\value{
A list of estimated parameters for the cosine model.
}
\description{
This function converts the estimated parameters from the NLS model fit to the
appropriate parameter space corresponding to the cosine model proposed by Balasse et al.
}
