% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBIC.R
\name{EBIC}
\alias{EBIC}
\title{Bayesian Information Criterion (BIC) for a partition.}
\usage{
EBIC(paths, partition, bandwidth)
}
\arguments{
\item{paths}{A list of data frames, where each frame contains the data for one individual. Every
data frame should have two columns with names 'distance' and 'oxygen'.}

\item{partition}{A list of vectors. Each element in the list is a vector of integers, corresponding
to individuals considered in one group.}

\item{bandwidth}{Denotes the order of the bandwidth that should be used in the estimation process.
bandwidth = k will mean that the bandwidth is n^k.}
}
\value{
Value of the extended BIC function for the partition.
}
\description{
This function calculates an extended version of BIC, which is computed using
a particular weighted average of the total residual sum of squares and the number of clusters.
}
