% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sits_cosine.R
\name{fit_cosine}
\alias{fit_cosine}
\title{Fit Cosine Model using Nonlinear Least Squares}
\usage{
fit_cosine(data, formula)
}
\arguments{
\item{data}{A data frame containing the stable isotope values for a single specimen.}

\item{formula}{Formula describing the independent and dependent variables, e.g. \code{oxygen~distance}. Will be converted to a cosine model.}
}
\value{
A fitted model object. See \code{\link[stats:nls]{stats::nls()}}.
}
\description{
Performs the updated nonlinear least squares (NLS) regression method for the cosine
model proposed by Balasse et al. The method calculates with the proposed initial values at first,
and then fits the NLS method as required.
}
\examples{
data(tsaghkahovit)
sits_cosine(tsaghkahovit[tsaghkahovit$id==6223,], oxygen~distance)

}
