% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDNAstemness}
\alias{getDNAstemness}
\title{Get stemness score from DNA methylation data}
\usage{
getDNAstemness(expr = NULL)
}
\arguments{
\item{expr}{a dataframe. Rowname is probe id. Colname is sample id.}
}
\value{
a named vector
}
\description{
Get stemness score from DNA methylation data
}
