% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getRNAstemness}
\alias{getRNAstemness}
\title{Get stemness score from transcriptome data}
\usage{
getRNAstemness(expr = NULL)
}
\arguments{
\item{expr}{a dataframe. Rowname is gene symbol id. Colname is sample id.}
}
\value{
a named vector
}
\description{
Get stemness score from transcriptome data
}
