% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFits.R
\name{makeFits}
\alias{makeFits}
\title{Prepare results for cosine model fit.}
\usage{
makeFits(paths)
}
\arguments{
\item{paths}{A list of data frames, where each frame contains the data for one individual. Every
data frame should have two columns with names 'distance' and 'oxygen'.}
}
\value{
A data frame with the estimated parameters in the model, birth seasonality estimate,
predicted/observed minimum/maximum for the oxygen isotope variable, mean squared error
and Pearson's R^2 corresponding to the model fit for every individual.
}
\description{
Performs the nonlinear least squares (NLS) regression method for the cosine
model, with the proposed initialization for all the parameters. It fits the NLS method
as required, and then computes different quantities for the birth seasonality estimates
corresponding to different individuals.
}
