% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCEM.R
\name{SCEM}
\alias{SCEM}
\title{Splitting-Coalescence-Estimation Method (SCEM) for archaeological time series.}
\usage{
SCEM(paths, bandwidth)
}
\arguments{
\item{paths}{A list of data frames, where each frame contains the data for one individual. There
should be two columns with names 'distance' and 'oxygen'.}

\item{bandwidth}{Denotes the order of the bandwidth that should be used in the splitting-coalescence
(SC) clustering algorithm. A value k will mean that the bandwidth used in the algorithm is n^k.}
}
\value{
A list containing the following components:

\item{results}{A data frame that has the individual information (ID, species, number of observations in the time series), cluster assignment, estimated period, delay and the birth seasonality estimate for every individual.}
\item{groups}{The groups formed by the clustering algorithm}
}
\description{
Performs the clustering algorithm SCEM on the bivariate time series data - where
one series is for the distance from the cementum-enamel junction, and the other series is for
the value of the oxygen-18 isotope at that distance - and returns the class-assignments and
birth seasobality estimates for all the individuals.
}
\examples{
armenia_split = split(armenia,f = armenia$ID)
results = SCEM(armenia_split,bandwidth = -0.33)
}
\references{
Chazin, Hannah, Soudeep Deb, Joshua Falk, and Arun Srinivasan. 2019. “New Statistical Approaches to Intra-Individual Isotopic Analysis and Modeling Birth Seasonality in Studies of Herd Animals.” Archaeometry 61 (2): 478–93.
}
