% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sineFit.R
\name{sineFit}
\alias{sineFit}
\title{Cosine model fitting with proposed initialization.}
\usage{
sineFit(data)
}
\arguments{
\item{data}{A data frame that contains the data for one individual. There should be two columns
with names 'distance' and 'oxygen'.}
}
\value{
A fitted model object from the nls function in R.
}
\description{
Performs the updated nonlinear least squares (NLS) regression method for the cosine
model proposed by Balasse et al. The method calculates with the proposed initial values at first,
and then fits the NLS method as required.
}
